feat: implement salary management use cases including creation and validation

+++ src/main/java/cl/prezdev/balancehub/application/usecases/fixedexpense/create/CreateFixedExpenseCommand.java
@@ -1,5 +1,7 @@
 package cl.prezdev.balancehub.application.usecases.fixedexpense.create;
 
+import java.math.BigDecimal;
+
 public record CreateFixedExpenseCommand(
-    String description, double amount
+    String description, BigDecimal amount
 ) {}
